Move pipelines to diart.pipelines. Add torchmetrics as a dependency
