+++ requirements.txt
@@ -9,6 +9,7 @@ pandas>=1.4.2
 torch>=1.12.1
 torchvision>=0.14.0
 torchaudio>=0.12.1,<1.0
+torchmetrics>=0.11.1
 pyannote.audio>=2.1.1
 pyannote.core>=4.5
 pyannote.database>=4.1.1
