+++ src/diart/models.py
@@ -1,4 +1,3 @@
-import time
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional, Text, Union, Callable, List, Any
