+++ src/diart/pipelines/__init__.py
@@ -1,0 +1,4 @@
+from .base import StreamingPipeline, StreamingConfig
+from .diarization import SpeakerDiarization, SpeakerDiarizationConfig
+from .transcription import Transcription, TranscriptionConfig
+from .voice import VoiceActivityDetection, VoiceActivityDetectionConfig
