+++ src/diart/pipelines/hparams.py
@@ -1,0 +1,24 @@
+from dataclasses import dataclass
+from typing import Text
+
+
+@dataclass
+class HyperParameter:
+    name: Text
+    low: float
+    high: float
+
+    @staticmethod
+    def from_name(name: Text) -> 'HyperParameter':
+        if name == "tau_active":
+            return TauActive
+        if name == "rho_update":
+            return RhoUpdate
+        if name == "delta_new":
+            return DeltaNew
+        raise ValueError(f"Hyper-parameter '{name}' not recognized")
+
+
+TauActive = HyperParameter("tau_active", low=0, high=1)
+RhoUpdate = HyperParameter("rho_update", low=0, high=1)
+DeltaNew = HyperParameter("delta_new", low=0, high=2)
